cookbook/docker/RAG-VLM/requirements.txt
gradio>=4.0.0
opencv-python>=4.8.0
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0

